🌱 split config god file
Move config loading, env parsing, defaults, validation, persistence, Kubernetes helpers, and supporting config structs into topical pkg/config files while keeping behavior unchanged.

Fixes #5643
