+++ changelog.d/changed-5643-config-split.md
@@ -1,0 +1,1 @@
+- Split `pkg/config/config.go` into topical load, env, defaults, validation, persistence, Kubernetes, and core type files without changing behavior.
