+++ src/pkg/config/env.go
@@ -1,0 +1,163 @@
+package config
+
+import (
+	"bufio"
+	"context"
+	"fmt"
+	"os"
+	"strings"
+
+	"github.com/kubestellar/hive/pkg/resolve"
+)
+
+func findConfigEnv(yamlPath string) string {
+	candidates := []string{
+		strings.TrimSuffix(yamlPath, "hive.yaml") + "config.env",
+		"/etc/hive/config.env",
+	}
+	for _, c := range candidates {
+		if _, err := os.Stat(c); err == nil {
+			return c
+		}
+	}
+	return ""
+}
+
+// ParseEnvFile reads a flat KEY=VALUE file (# comments, blank lines skipped).
+func ParseEnvFile(path string) (map[string]string, error) {
+	f, err := os.Open(path)
+	if err != nil {
+		return nil, err
+	}
+	defer func() { _ = f.Close() }() // read-only fd; nothing to lose on close error
+
+	result := make(map[string]string)
+	scanner := bufio.NewScanner(f)
+	for scanner.Scan() {
+		line := strings.TrimSpace(scanner.Text())
+		if line == "" || strings.HasPrefix(line, "#") {
+			continue
+		}
+		key, val, ok := strings.Cut(line, "=")
+		if !ok {
+			continue
+		}
+		key = strings.TrimSpace(key)
+		val = strings.TrimSpace(val)
+		val = strings.Trim(val, `"'`)
+		result[key] = val
+	}
+	return result, scanner.Err()
+}
+
+// applyConfigEnv merges flat KEY=VALUE overrides into the loaded config.
+func (c *Config) applyConfigEnv(path string) error {
+	env, err := ParseEnvFile(path)
+	if err != nil {
+		return err
+	}
+
+	if v, ok := env["PROJECT_ORG"]; ok {
+		c.Project.Org = v
+	}
+	if v, ok := env["PROJECT_REPOS"]; ok {
+		c.Project.Repos = strings.Fields(v)
+	}
+	if v, ok := env["PROJECT_AI_AUTHOR"]; ok {
+		c.Project.AIAuthor = v
+	}
+	if v, ok := env["PROJECT_PRIMARY_REPO"]; ok {
+		c.Project.PrimaryRepo = v
+	}
+	if v, ok := env["PROJECT_OPEN_PRS"]; ok {
+		b := v == "true" || v == "1" || v == "yes"
+		c.Project.OpenPRs = &b
+	}
+	if v, ok := env["AGENTS_ENABLED"]; ok {
+		for _, name := range strings.Fields(v) {
+			if agent, exists := c.Agents[name]; exists {
+				agent.Enabled = true
+				c.Agents[name] = agent
+			}
+		}
+	}
+	if v, ok := env["DASHBOARD_PORT"]; ok {
+		var port int
+		if _, err := fmt.Sscanf(v, "%d", &port); err == nil && port > 0 {
+			c.Dashboard.Port = port
+		}
+	}
+	if v, ok := env["DASHBOARD_AUTH_TOKEN"]; ok {
+		c.Dashboard.AuthToken = v
+	}
+	if c.Dashboard.AuthToken == "" {
+		if v, ok := env["HIVE_DASHBOARD_TOKEN"]; ok {
+			c.Dashboard.AuthToken = v
+		}
+	}
+
+	return nil
+}
+
+func (c *Config) applyBootstrapEnv() {
+	if repo := os.Getenv("HIVE_REPO"); repo != "" {
+		parts := strings.SplitN(repo, "/", 2)
+		if len(parts) == 2 && parts[0] != "" && parts[1] != "" {
+			if c.Project.Org == "" {
+				c.Project.Org = parts[0]
+			}
+			if len(c.Project.Repos) == 0 {
+				c.Project.Repos = []string{parts[1]}
+			}
+			if c.Project.PrimaryRepo == "" {
+				c.Project.PrimaryRepo = parts[1]
+			}
+		}
+	}
+	// K8s deployments pass the auth token as an OS env var from a Secret.
+	// applyConfigEnv only reads file-based config.env, so without this
+	// the token is silently ignored and the dashboard is unauthenticated.
+	if c.Dashboard.AuthToken == "" {
+		if v := os.Getenv("DASHBOARD_AUTH_TOKEN"); v != "" {
+			c.Dashboard.AuthToken = v
+		}
+	}
+	if c.Dashboard.AuthToken == "" {
+		if v := os.Getenv("HIVE_DASHBOARD_TOKEN"); v != "" {
+			c.Dashboard.AuthToken = v
+		}
+	}
+	// K8s-provisioned spokes receive their per-hive authorized GitHub users as a
+	// comma-separated env var (owner first). This is what lets a direct-route
+	// spoke reject unauthorized device-flow logins without the hub proxy.
+	if len(c.Dashboard.AuthorizedUsers) == 0 {
+		if v := os.Getenv("HIVE_AUTHORIZED_USERS"); v != "" {
+			c.Dashboard.AuthorizedUsers = parseAuthorizedUsers(v)
+		}
+	}
+}
+
+// parseAuthorizedUsers splits a comma-separated authorized-users list, trimming
+// whitespace and dropping empty entries. Order is preserved so the first entry
+// remains the owner.
+func parseAuthorizedUsers(v string) []string {
+	parts := strings.Split(v, ",")
+	out := make([]string, 0, len(parts))
+	for _, p := range parts {
+		if u := strings.TrimSpace(p); u != "" {
+			out = append(out, u)
+		}
+	}
+	return out
+}
+
+// expandEnvVars substitutes ${VAR} references in the raw config text. It runs
+// BEFORE the YAML is parsed, so to honor an operator `variables:` block it
+// first bootstrap-parses just that block from the same text, builds a
+// config-scoped resolve.Registry, and delegates. With no `variables:` block the
+// registry is env-only and the result is byte-identical to the legacy behavior
+// (${NAME} -> os.LookupEnv(NAME), unset left literal).
+func expandEnvVars(s string) string {
+	reg := configRegistryFromText(s)
+	return reg.Expand(context.Background(), s, resolve.ScopeConfig, nil)
+}
