+++ src/pkg/config/kubernetes.go
@@ -1,0 +1,48 @@
+package config
+
+import "os"
+
+var saTokenFile = "/var/run/secrets/kubernetes.io/serviceaccount/token"
+
+// SetSATokenFileForTest points IsKubernetesPod's serviceaccount-token probe
+// at path and returns a restore func. Out-of-package tests that need the
+// non-Kubernetes branch call this with a non-existent path (alongside
+// clearing KUBERNETES_SERVICE_HOST) so they stay hermetic on hosts that
+// really are pods — in-cluster CI runners and dev hives.
+func SetSATokenFileForTest(path string) func() {
+	orig := saTokenFile
+	saTokenFile = path
+	return func() { saTokenFile = orig }
+}
+
+// IsKubernetesPod reports whether the process is running inside a
+// Kubernetes pod (mirrors the entrypoint's IS_KUBERNETES detection).
+func IsKubernetesPod() bool {
+	if os.Getenv("KUBERNETES_SERVICE_HOST") != "" {
+		return true
+	}
+	_, err := os.Stat(saTokenFile)
+	return err == nil
+}
+
+// saveDashboardOverlay writes the secret-free PVC overlay in Kubernetes
+// mode. Failures are logged, never fatal — but they ARE returned (#3961):
+// when the primary config path is unwritable (read-only ConfigMap mount)
+// the overlay and the runtime config are the only layers that survive a pod
+// restart, so saveLocked needs to know whether this write landed before it
+// can report the save as durable. Outside Kubernetes it returns nil (the
+// overlay is not part of the boot path there).
+//
+// The write MUST be atomic (temp file + rename), unlike saveLocked()'s
+// inode-preserving write to the bind-mounted primary config. DashboardOverlayFile
+// lives on the PVC (not a bind mount), so rename is safe here, and it is the
+// only way to avoid a truncated/partial overlay if the pod is killed mid-write
+// (a redeploy sends SIGTERM/SIGKILL at an arbitrary instant). A truncate-in-place
+// write (os.WriteFile) can leave the file cut off partway through — GitHubConfig
+// marshals AFTER Agents/Project in the Config struct field order (see the
+// struct tags above), so a truncated overlay can silently keep valid
+// project/agents blocks while losing app_id/installation_id/key_file entirely.
+// The entrypoint's merge script only sanity-checks project.org and agents
+// before trusting the overlay wholesale, so that truncated-but-plausible file
+// would pass the guard and revert a dashboard-installed GitHub App to the
+// placeholder ConfigMap seed on the next restart — exactly the durability bug
