+++ src/pkg/config/policies_config.go
@@ -1,0 +1,11 @@
+package config
+
+import "time"
+
+type PoliciesConfig struct {
+	Repo         string        `yaml:"repo"`
+	Branch       string        `yaml:"branch"`
+	Path         string        `yaml:"path"`
+	PollInterval time.Duration `yaml:"poll_interval"`
+	LocalDir     string        `yaml:"local_dir"`
+}
